requirement.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.2
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gym==0.15.3
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyglet==1.3.2
pyparsing==2.4.4
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.1
six==1.13.0
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
